fix warning

+++ textaddr.c
@@ -4,7 +4,7 @@ extern long etext;
 
 int main(int argc, char *argv[])
 {
-	printf("%lu\n", &etext);
+	printf("%lu\n", (unsigned long)&etext);
 	return 0;
 }
